determines selectMode implicitly
