+++ packages/payload/src/utilities/stripUnselectedFields.ts
@@ -1,9 +1,15 @@
 import type { Data } from '../admin/types.js'
-import type { Block, Field, TabAsField } from '../fields/config/types.js'
+import type { Field, TabAsField } from '../fields/config/types.js'
 import type { SelectMode, SelectType } from '../types/index.js'
 
 import { fieldAffectsData } from '../fields/config/types.js'
 
+/**
+ * This is used for the Select API to strip out fields that are not selected.
+ * It will mutate the given data object and determine if your recursive function should continue to run.
+ * It is used within the `afterRead` hook and within form state.
+ * @returns boolean - whether or not the recursive function should continue
+ */
 export const stripUnselectedFields = ({
   field,
   select,
@@ -34,48 +40,4 @@ export const stripUnselectedFields = ({
   }
 
   return shouldContinue
-}
-
-export const handleBlocksSelect = ({
-  block,
-  select,
-  selectMode,
-}: {
-  block: Block
-  select: boolean | SelectType
-  selectMode: SelectMode
-}): { blockSelect: SelectType; blockSelectMode: SelectMode } => {
-  let blocksSelect = select
-  let blockSelectMode = selectMode
-
-  if (typeof select === 'object') {
-    blocksSelect = {
-      ...select,
-    }
-
-    // sanitize blocks: {cta: false} to blocks: {cta: {id: true, blockType: true}}
-    if (selectMode === 'exclude' && blocksSelect[block.slug] === false) {
-      blockSelectMode = 'include'
-
-      select[block.slug] = {
-        id: true,
-        blockType: true,
-      }
-    } else if (selectMode === 'include') {
-      if (!blocksSelect[block.slug]) {
-        blocksSelect[block.slug] = {}
-      }
-
-      if (typeof blocksSelect[block.slug] === 'object') {
-        blocksSelect[block.slug] = {
-          ...(blocksSelect[block.slug] as object),
-        }
-
-        blocksSelect[block.slug]['id'] = true
-        blocksSelect[block.slug]['blockType'] = true
-      }
-    }
-  }
-
-  return { blockSelect: blocksSelect?.[block.slug], blockSelectMode }
 }
